+++ src/views/library/page/read-note/index.tsx
@@ -10,7 +10,7 @@ import read_note_save_btn from '/src/assets/button/read-note/read-note-save-btn.
 import quotation_arrow from '/src/assets/button/read-note-edit/quotation-arrow.svg';
 import impression_icon from '/src/assets/button/read-note-edit/impression-icon.svg';
 
-import LibraryRegistration from '../../../home/components/SearchLibraryRegistration';
+import LibraryRegistration from '../../../lounge/components/book-info/libraryRegistration';
 import DeleteBtn from '../../../../components/delete-modal/DeleteModal';
 
 const ReadNotePage = () => {
